[refactor] - 커뮤니티 페이지 스와이프 가능하게 변경
